Изменение структуры view.
