+++ src/view/board-view/board-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import {boardTemplate} from './board.tpl';
+
+export default class BoardView {
+  getTemplate() {
+    return boardTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
