+++ src/view/board-view/board.tpl.js
@@ -1,0 +1,1 @@
+export const boardTemplate = () => '<section class="films"></section>';
