+++ src/view/list-title-view/list-title-tpl.js
@@ -1,0 +1,1 @@
+export const listTitleTemplate = () => '<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>';
