+++ src/view/list-view/list-tpl.js
@@ -1,0 +1,5 @@
+export const listTemplate = () => (
+  `<section class="films-list">
+      <div class="films-list__container"></div
+   </section>`
+);
