+++ src/view/more-button-view/more-button-tpl.js
@@ -1,0 +1,1 @@
+export const moreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
