+++ src/view/more-button-view/more-button-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import {moreButtonTemplate} from './more-button-tpl';
+
+export default class MoreButtonView {
+  getTemplate() {
+    return moreButtonTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
