+++ src/view/navigation-view/navigation-tpl.js
@@ -1,6 +1,4 @@
-import {createElement} from '../render.js';
-
-const navigationTemplate = () => (
+export const navigationTemplate = () => (
   `<nav class="main-navigation">
     <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
     <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
@@ -8,21 +6,3 @@ const navigationTemplate = () => (
     <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
   </nav>`
 );
-
-export default class NavigationView {
-  getTemplate() {
-    return navigationTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
