+++ src/view/navigation-view/navigation-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import {navigationTemplate} from './navigation-tpl';
+
+export default class NavigationView {
+  getTemplate() {
+    return navigationTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
