+++ src/view/statistics-view/statistics-tpl.js
@@ -1,0 +1,1 @@
+export const statisticsTemplate = () => '<p>130 291 movies inside</p>';
