+++ src/view/user-name-view/user-name-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import {userNameTemplate} from './user-name-tpl';
+
+export default class UserNameView {
+  getTemplate() {
+    return userNameTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
